Change SuiteUser ID type to Integer. Closes #1823.
While CSM's User#userId property is typed as a long, CSM's schema does not allow
for greater than 32-bit signed integers as ID values. PSC's references to what were
originally CSM-sourced IDs followed this pattern. Using the Integer type for the
SuiteUser ID makes it clear to implementors that only 32-bit signed integers are
usable with PSC.

+++ authorization/plugin-api/src/main/java/gov/nih/nci/cabig/ctms/suite/authorization/plugin/SuiteUser.java
@@ -24,7 +24,7 @@ import java.util.TreeMap;
  * @author Rhett Sutphin
  */
 public class SuiteUser {
-    private Long id;
+    private Integer id;
     private String username, firstName, lastName, emailAddress;
     private Date accountEndDate;
 
@@ -70,11 +70,11 @@ public class SuiteUser {
     /**
      * A unique numeric ID for the user. Required.
      */
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -177,7 +177,7 @@ public class SuiteUser {
             return this;
         }
 
-        public Builder id(long id) {
+        public Builder id(int id) {
             user.setId(id);
             return this;
         }
